pyproject.toml
[tool.poetry]
name = "globant-challenge"
version = "0.1.0"
description = ""
authors = ["PabloBrenesA97 <josepablobrenesalfaro@gmail.com>"]
readme = "README.md"
packages = [{include = "globant_challenge"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
python-multipart = "^0.0.6"
python-magic = "^0.4.27"
boto3 = "^1.26.136"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
